refactor: make fontSize an int
It's set as an int upstream, assume it's corrupt
otherwise

+++ AnkiDroid/src/main/java/com/ichi2/anki/cardviewer/TypeAnswer.kt
@@ -114,7 +114,7 @@ class TypeAnswer(
                     correct = contentForCloze(correct!!, clozeIdx)
                 }
                 font = fld.font
-                size = fld.fontSize.toInt()
+                size = fld.fontSize
                 languageHint = fld.languageHint
                 break
             }
